+++ pkg/conf/computed.go
@@ -298,3 +298,8 @@ func IsBuiltinSignupAllowed() bool {
 	}
 	return false
 }
+
+// DisableUsernameChanges returns true if DisableUsernameChanges is true in the site configuration.
+func DisableUsernameChanges() bool {
+	return Get().DisableUsernameChanges
+}